backend/ai_rally/requirements.txt
ultralytics>=8.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
pygame>=2.5.0
websockets>=12.0
numpy>=1.24.0
roboflow>=1.0.0
pyyaml>=6.0
